split the Admin contract
